docker fix
we had a problem with react and docker where we had to rebuilt everytime before being able to see our modification. we fixed  the issue with modifying the yml with a dev and adding more context and mounts, we also had to force implement the vite files by providing the right port.